Fix warnings for Alpha CPU.
Whenever "%llu" is used as a format, cast the argument to long long.  On
Alpha, u_int64_t is not equivalent to unsigned long long.

+++ trunk/ath/if_ath.c
@@ -3751,7 +3751,7 @@ ath_beacon_alloc(struct ath_softc *sc, struct ieee80211_node *ni)
 		DPRINTF(sc, ATH_DEBUG_BEACON,
 			"%s: %s beacons, bslot %d intval %u tsfadjust %llu\n",
 			__func__, sc->sc_stagbeacons ? "stagger" : "burst",
-			avp->av_bslot, ni->ni_intval, tsfadjust);
+			avp->av_bslot, ni->ni_intval, (long long) tsfadjust);
 
 		wh = (struct ieee80211_frame *) skb->data;
 		memcpy(&wh[1], &tsfadjust, sizeof(tsfadjust));
@@ -4066,7 +4066,7 @@ ath_beacon_send(struct ath_softc *sc, int *needmark)
 		vap = sc->sc_bslot[(slot + 1) % ATH_BCBUF];
 		DPRINTF(sc, ATH_DEBUG_BEACON_PROC,
 			"%s: slot %d [tsf %llu tsftu %u intval %u] vap %p\n",
-			__func__, slot, tsf, tsftu, ic->ic_lintval, vap);
+			__func__, slot, (long long) tsf, tsftu, ic->ic_lintval, vap);
 		bfaddr = 0;
 		if (vap != NULL) {
 			bf = ath_beacon_generate(sc, vap, needmark);
@@ -4417,7 +4417,7 @@ ath_beacon_config(struct ath_softc *sc, struct ieee80211vap *vap)
 		DPRINTF(sc, ATH_DEBUG_BEACON, 
 			"%s: tsf %llu tsf:tu %u intval %u nexttbtt %u dtim %u nextdtim %u bmiss %u sleep %u cfp:period %u maxdur %u next %u timoffset %u\n"
 			, __func__
-			, tsf, tsftu
+			, (long long) tsf, tsftu
 			, bs.bs_intval
 			, bs.bs_nexttbtt
 			, bs.bs_dtimperiod
@@ -5317,8 +5317,8 @@ ath_recv_mgmt(struct ieee80211_node *ni, struct sk_buff *skb,
 			if (le64_to_cpu(ni->ni_tstamp.tsf) >= tsf) {
 				DPRINTF(sc, ATH_DEBUG_STATE,
 					"ibss merge, rstamp %u tsf %llu "
-					"tstamp %llu\n", rstamp, tsf,
-					ni->ni_tstamp.tsf);
+					"tstamp %llu\n", rstamp, (long long) tsf,
+					(long long) ni->ni_tstamp.tsf);
 				(void) ieee80211_ibss_merge(ni);
 			}
 		}
